move off index manager util for configure index test, move collection error tests into the CollectionTest file
