+++ src/integration/java/io/pinecone/integration/controlPlane/pod/CollectionTest.java
@@ -2,6 +2,7 @@ package io.pinecone.integration.controlPlane.pod;
 
 import io.pinecone.clients.Pinecone;
 import io.pinecone.configs.*;
+import io.pinecone.exceptions.PineconeException;
 import io.pinecone.helpers.RandomStringBuilder;
 import io.pinecone.proto.*;
 import org.junit.jupiter.api.AfterAll;
@@ -11,9 +12,7 @@ import org.openapitools.client.model.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.util.*;
 
 import static io.pinecone.helpers.IndexManager.createNewIndexAndConnect;
 import static io.pinecone.helpers.IndexManager.waitUntilIndexIsReady;
@@ -187,5 +186,94 @@ public class CollectionTest {
         assertEquals(newIndex.getMetric(), targetMetric);
 
         dataPlaneConnection.close();
+    }
+
+    @Test
+    public void testCreateCollectionFromInvalidIndex() {
+        try {
+            CreateCollectionRequest createCollectionRequest = new CreateCollectionRequest().name(RandomStringBuilder.build("coll1", 8)).source("invalid-index");
+            controlPlaneClient.createCollection(createCollectionRequest);
+        } catch (PineconeException exception) {
+            logger.info("Exception: " + exception.getMessage());
+            assertTrue(exception.getMessage().contains("Resource invalid-index not found"));
+        }
+    }
+    @Test
+    public void testIndexFromNonExistentCollection() {
+        try {
+            CreateIndexRequestSpecPod podSpec = new CreateIndexRequestSpecPod().environment(environment).sourceCollection("non-existent-collection");
+            CreateIndexRequestSpec spec = new CreateIndexRequestSpec().pod(podSpec);
+            CreateIndexRequest newCreateIndexRequest = new CreateIndexRequest().name(RandomStringBuilder.build("from-nonexistent-coll", 8)).dimension(dimension).metric(IndexMetric.COSINE).spec(spec);
+            controlPlaneClient.createIndex(newCreateIndexRequest);
+        } catch (PineconeException exception) {
+            logger.info("Exception: " + exception.getMessage());
+            assertTrue(exception.getMessage().contains("Resource non-existent-collection not found"));
+        }
+    }
+
+    @Test
+    public void testCreateIndexInMismatchedEnvironment() {
+        try {
+            List<String> environments = new LinkedList<>(Arrays.asList(
+                    "eastus-azure",
+                    "eu-west4-gcp",
+                    "northamerica-northeast1-gcp",
+                    "us-central1-gcp",
+                    "us-west4-gcp",
+                    "asia-southeast1-gcp",
+                    "us-east-1-aws",
+                    "asia-northeast1-gcp",
+                    "eu-west1-gcp",
+                    "us-east1-gcp",
+                    "us-east4-gcp",
+                    "us-west1-gcp"
+            ));
+            environments.remove(collection.getEnvironment());
+            String mismatchedEnv = environments.get(new Random().nextInt(environments.size()));
+
+            CreateIndexRequestSpecPod podSpec = new CreateIndexRequestSpecPod().sourceCollection(collection.getName()).environment(mismatchedEnv);
+            CreateIndexRequestSpec spec = new CreateIndexRequestSpec().pod(podSpec);
+            CreateIndexRequest createIndexRequest = new CreateIndexRequest().name(RandomStringBuilder.build("from-coll", 8)).dimension(dimension).metric(IndexMetric.COSINE).spec(spec);
+            controlPlaneClient.createIndex(createIndexRequest);
+        } catch (PineconeException exception) {
+            logger.info("Exception: " + exception.getMessage());
+            assertTrue(exception.getMessage().contains("Source collection must be in the same environment as the index"));
+        }
+    }
+
+    @Test
+    @Disabled("Bug reported in #global-cps")
+    public void testCreateIndexWithMismatchedDimension() {
+        try {
+            CreateIndexRequestSpecPod podSpec = new CreateIndexRequestSpecPod().sourceCollection(collection.getName()).environment(collection.getEnvironment());
+            CreateIndexRequestSpec spec = new CreateIndexRequestSpec().pod(podSpec);
+            CreateIndexRequest createIndexRequest = new CreateIndexRequest().name(RandomStringBuilder.build("from-coll", 8)).dimension(dimension + 1).metric(IndexMetric.COSINE).spec(spec);
+            controlPlaneClient.createIndex(createIndexRequest);
+        } catch (PineconeException exception) {
+            logger.info("Exception: " + exception.getMessage());
+            assertTrue(exception.getMessage().contains("Index and collection must have the same dimension"));
+        }
+    }
+
+    @Test
+    public void testCreateCollectionFromNotReadyIndex() throws InterruptedException {
+        String notReadyIndexName = RandomStringBuilder.build("from-coll", 8);
+        String newCollectionName = RandomStringBuilder.build("coll-", 8);
+        try {
+            CreateIndexRequestSpecPod specPod = new CreateIndexRequestSpecPod().pods(1).podType("p1.x1").replicas(1).environment(environment);
+            CreateIndexRequestSpec spec = new CreateIndexRequestSpec().pod(specPod);
+            CreateIndexRequest createIndexRequest = new CreateIndexRequest().name(notReadyIndexName).dimension(dimension).metric(IndexMetric.COSINE).spec(spec);
+            controlPlaneClient.createIndex(createIndexRequest);
+            indexes.add(notReadyIndexName);
+
+            createCollection(controlPlaneClient, newCollectionName, notReadyIndexName, true);
+        } catch (PineconeException exception) {
+            logger.info("Exception: " + exception.getMessage());
+            assert (exception.getMessage().contains("Source index is not ready"));
+        } finally {
+            Thread.sleep(2500);
+            // clean up index and collection
+            controlPlaneClient.deleteIndex(notReadyIndexName);
+        }
     }
 }
